refactor(shared): move Inflection*, update imports (no behavor change)
