+++ src/main/java/com/annepolis/lexiconmeum/shared/model/Inflection.java
@@ -1,4 +1,4 @@
-package com.annepolis.lexiconmeum.lexeme.detail;
+package com.annepolis.lexiconmeum.shared.model;
 
 public interface Inflection {
