+++ src/main/java/com/annepolis/lexiconmeum/shared/model/Lexeme.java
@@ -1,6 +1,5 @@
 package com.annepolis.lexiconmeum.shared.model;
 
-import com.annepolis.lexiconmeum.lexeme.detail.Inflection;
 import com.annepolis.lexiconmeum.shared.model.grammar.GrammaticalPosition;
 import com.annepolis.lexiconmeum.shared.model.grammar.InflectionClass;
 
@@ -37,7 +36,7 @@ public class Lexeme {
 
     public String getLemma() { return lemma; }
 
-    public GrammaticalPosition getPosition() {
+    public GrammaticalPosition getGrammaticalPosition() {
         return position;
     }
