+++ src/main/java/com/annepolis/lexiconmeum/shared/model/LexemeBuilder.java
@@ -1,7 +1,5 @@
 package com.annepolis.lexiconmeum.shared.model;
 
-import com.annepolis.lexiconmeum.lexeme.detail.Inflection;
-import com.annepolis.lexiconmeum.lexeme.detail.InflectionKey;
 import com.annepolis.lexiconmeum.shared.model.grammar.GrammaticalGender;
 import com.annepolis.lexiconmeum.shared.model.grammar.GrammaticalPosition;
 import com.annepolis.lexiconmeum.shared.model.grammar.InflectionClass;
