+++ src/main/java/com/annepolis/lexiconmeum/textsearch/TextSearchTrieIndexService.java
@@ -1,6 +1,5 @@
 package com.annepolis.lexiconmeum.textsearch;
 
-import com.annepolis.lexiconmeum.lexeme.detail.Inflection;
 import com.annepolis.lexiconmeum.shared.LexemeSink;
 import com.annepolis.lexiconmeum.shared.model.Lexeme;
 import com.github.benmanes.caffeine.cache.Cache;
