+++ src/test/java/com/annepolis/lexiconmeum/lexeme/detail/adjective/LexemeAgreementDetailMapperTest.java
@@ -1,8 +1,8 @@
 package com.annepolis.lexiconmeum.lexeme.detail.adjective;
 
 import com.annepolis.lexiconmeum.TestUtil;
-import com.annepolis.lexiconmeum.lexeme.detail.InflectionKey;
 import com.annepolis.lexiconmeum.lexeme.detail.LexemeDetailResponse;
+import com.annepolis.lexiconmeum.shared.model.InflectionKey;
 import com.annepolis.lexiconmeum.shared.model.Lexeme;
 import org.junit.jupiter.api.Test;
