+++ src/test/java/com/annepolis/lexiconmeum/lexeme/detail/verb/InflectionKeyTest.java
@@ -1,7 +1,7 @@
 package com.annepolis.lexiconmeum.lexeme.detail.verb;
 
 import com.annepolis.lexiconmeum.TestUtil;
-import com.annepolis.lexiconmeum.lexeme.detail.InflectionKey;
+import com.annepolis.lexiconmeum.shared.model.InflectionKey;
 import com.annepolis.lexiconmeum.shared.model.Lexeme;
 import com.annepolis.lexiconmeum.shared.model.grammar.*;
 import org.junit.jupiter.api.Assertions;
