+++ src/test/java/com/annepolis/lexiconmeum/lexeme/detail/verb/LexemeConjugationDetailMapperTest.java
@@ -1,9 +1,9 @@
 package com.annepolis.lexiconmeum.lexeme.detail.verb;
 
 import com.annepolis.lexiconmeum.TestUtil;
-import com.annepolis.lexiconmeum.lexeme.detail.InflectionKey;
 import com.annepolis.lexiconmeum.lexeme.detail.LexemeDetailResponse;
 import com.annepolis.lexiconmeum.lexeme.detail.LexemeInflectionMapper;
+import com.annepolis.lexiconmeum.shared.model.InflectionKey;
 import com.annepolis.lexiconmeum.shared.model.Lexeme;
 import org.junit.jupiter.api.Test;
